size_t fixed

+++ circle_1/libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: hyeslim <hyeslim@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 16:31:58 by hyeslim           #+#    #+#             */
-/*   Updated: 2022/07/08 12:54:55 by hyeslim          ###   ########.fr       */
+/*   Updated: 2022/07/11 15:52:16 by hyeslim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	char	*dest;
-	int		i;
+	char		*dest;
+	size_t		i;
 
 	i = 0;
 	dest = (char *)malloc(sizeof(char) * len);
 	if (!dest)
 		return (0);
-	while (i + 1 < (int)len)
+	while (i + 1 < len)
 	{
 		dest[i] = s[start + i];
 		i++;
